+++ .editorconfig
@@ -6,7 +6,7 @@ root = true
 # Python specific settings, complying with PEP 8 style guide
 [*.py]
 indent_size = 4
-max_line_length = 80
+max_line_length = 100
 
 # Markdown files
 [*.md]
@@ -21,7 +21,7 @@ indent_size = 2
 indent_size = 2
 
 # YAML files
-[*.yml]
+[*.{yaml,yml}]
 indent_size = 2
 
 # JSON files
